refactor v2 eval collection sequence
